requirements.txt
fastapi==0.81.0
SQLAlchemy==1.4.40
pydantic==1.10.1
pyhumps==3.7.3
uvicorn==0.18.3
pika==1.3.0
aio-pika==8.2.0
alembic==1.8.1
psycopg2==2.9.3
